+++ src/adaptors/cyberconnect.ts
@@ -3,16 +3,16 @@ import { gqlFetcher } from '../utils';
 export default async function getCyberconnectData(address: string) {
   const data = await gqlFetcher(
     'https://api.cybertino.io/connect/',
-    `{
-        identity(address: "${address.toString()}") {
-            displayName
-            address
-            followingCount
-            followerCount
-            social {
-                twitter
-            }
+    `query {
+      identity(address: "${address.toString()}") {
+        displayName
+        address
+        followingCount
+        followerCount
+        social {
+          twitter
         }
+      }
     }`
   );