+++ src/adaptors/index.ts
@@ -23,10 +23,10 @@ export { default as getPolygonData } from './polygon';
 export { default as getProjectGalaxyData } from './projectgalaxy';
 export { default as checkPoH } from './poh';
 export { default as getRabbitholeData } from './rabbithole';
-// Rarible
+export { default as getRaribleData } from './rarible';
 export { default as getRss3Data } from './rss3';
 export { default as getShowtimeData } from './showtime';
-// Superrare
-// sybil
+export { default as getSuperrareData } from './superrare';
+export { getAllSybilData, getSybilData } from './uniswap';
 export { default as resolveUnstoppableDomains } from './unstoppable';
-// zora
+export { default as getZoraData } from './zora';